requirements-dev.txt
-i http://pypi.nowoczesnapolska.org.pl/simple

django-debug-toolbar
fnpdjango
Fabric
